+++ plugins/generic-table/protocol_table_dirlisting.c
@@ -263,7 +263,7 @@ callback_lws_table_dirlisting(struct lws *wsi, enum lws_callback_reasons reason,
 				q += strlen(q);
 			} else {
 				n = q1 - q;
-				if (n > sizeof(s) - 1)
+				if (n > (int)sizeof(s) - 1)
 					n = sizeof(s) - 1;
 				if (first) {
 					strcpy(s1, "/");
@@ -273,7 +273,7 @@ callback_lws_table_dirlisting(struct lws *wsi, enum lws_callback_reasons reason,
 					s[n] = '\0';
 
 					n = q1 - pss->reldir;
-					if (n > sizeof(s1) - 1)
+					if (n > (int)sizeof(s1) - 1)
 						n = sizeof(s1) - 1;
 					strncpy(s1, pss->reldir, n);
 					s1[n] = '\0';
