+++ plugins/ssh-base/sshd.c
@@ -1332,6 +1332,7 @@ again:
 			if (!m) {
 				/* the decrypted sig is in ASN1 format */
 				m = 0;
+
 				while (m < olen) {
 					/* sig payload */
 					if (otmp[m] == 0x04 &&
@@ -2271,7 +2272,7 @@ lws_callback_raw_sshd(struct lws *wsi, enum lws_callback_reasons reason,
 			 *    string    public key blob from the request
       			 */
 			n = 74 + pss->ua->pubkey_len;
-			if (n > sizeof(buf) - LWS_PRE) {
+			if (n > (int)sizeof(buf) - LWS_PRE) {
 				lwsl_notice("pubkey too large\n");
 				goto bail;
 			}
